Changes for MacOS and simpler Linux
This adds what we need to be able to run builds with MacOS runners. This also makes the overall flow simpler by containing which runner we should use within the object we get from the set-matrix step itself so we can use a single build job.

Additionally, along with changes to vanagon and build repos, we now
build a single binary for all supported MacOS versions, and call the
platform string 'macos-all-<arch>'.

Finally, this adds an input to allow overriding the vanagon branch to
more easily test changes.