Add /etc/passwd to Cygwin environment
Because we now use the Puppet library to look up the version, Puppet gets confused when running inside a Cygwin shell. The check for determining if we're in Windows returns false because within the context of Cygwin, it looks like it's not. But the check for a POSIX environment also fails because Cygwin does not provide an /etc/passwd file. This lays down that file so when we load Puppet libraries, it doesn't throw up. It should not affect any other part of the build process.