Use Beaker suite name as identifier for test suite defaults
The project-name input used to be synonymous with the actual repository
containing the test suite to execute. But since the separate
openvox-agent repository was deprecated and pushed into
openvox/packaging, that no longer applies, and we have two test suites
inside the openvox repository, openvox/acceptance (the openvox test
suite) and openvox/packaging/acceptance (the openvox-agent test suite)
that we need to be able to execute.

This commit changes project-name to suite-name, and tracks the actual
repository as a default selected with the rest of the defaults
associated with suite-name.