Install openvoxdb-termini for acceptance based on suite
The acceptance-pipelines workflow sets openvoxdb vars because it runs
all the suites, so the openvox suite, which installs openvox-server, was
able to install openvoxdb-termini because the openvoxdb versions vars
happened to be set. But the openvoxproject/openvox acceptance pipeline
was failing to install openvoxdb-termini because it only gets openvox
and openvox-server version vars from its inputs.

This change determines whether to install the openvoxdb termini based on
the suite being run. Only the openvox-server and openvoxdb defaults have
install-termini true now.